initialize更新

+++ initialize.zsh
@@ -1,6 +1,23 @@
-ln -fs "$HOME/dotfiles/zsh/" "$HOME/.config/zsh"
-ln -fs "$HOME/dotfiles/zshrc" "$HOME/.zshrc"
-ln -fs "$HOME/dotfiles/zsh/alias.zsh" "$HOME/zalias"
-ln -fs "$HOME/dotfiles/git/" "$HOME/.config/git"
-ln -fs "$HOME/dotfiles/nvim/" "$HOME/.config/nvim"
-ln -fs "$HOME/dotfiles/brew/Brewfile" "$HOME/Brewfile"
+if [ ! -e "$HOME/.config/zsh" ]; then
+  ln -fs "$HOME/dotfiles/zsh/" "$HOME/.config/zsh"
+fi
+
+if [ ! -e "$HOME/.zshrc" ]; then
+  ln -fs "$HOME/dotfiles/zshrc" "$HOME/.zshrc"
+fi
+
+if [ ! -e "$HOME/zalias" ]; then
+  ln -fs "$HOME/dotfiles/zsh/alias.zsh" "$HOME/zalias"
+fi
+
+if [ ! -e "$HOME/.config/git" ]; then
+  ln -fs "$HOME/dotfiles/git/" "$HOME/.config/git"
+fi
+
+if [ ! -e "$HOME/.config/nvim" ]; then
+  ln -fs "$HOME/dotfiles/nvim/" "$HOME/.config/nvim"
+fi
+
+if [ ! -e "$HOME/Brewfile" ]; then
+  ln -fs "$HOME/dotfiles/brew/Brewfile" "$HOME/Brewfile"
+fi
